coalescing fully implemented. More bug testing required, but looks very promising so far.

+++ Assignment_2/C/M_Init.h
@@ -17,7 +17,7 @@ typedef struct memStruct
 }memStruct;
 
 void* freeList;
-int freeListSize;
+unsigned long freeListSize;
 void* magicNumber;
 memStruct* currentBlock;
